refactor: storage api 함수 공통으로 분리
